Added new MemoryCache related classes for proxy and service ticket managers.  This also required a singleton manager for the MemoryCache itself.
Change code that wires up the default in-memory proxy and service ticket managers in CasAuthentication.cs.

The .NET 2/3.5 assemblies will load the traditional (System.Web.Caching) proxy and service ticket managers that were already in the project.  The .NET 4.x assemblies will load the new MemoryCache (System.Runtime.Caching) proxy and service ticket managers.  There is no need to change the default configuration values for the proxy and service ticket managers in the web.config as this is a behind-the-scenes change.